backend/requirements.txt
lxml==4.0.0
algoliasearch==1.15.2
requests==2.18.4
beautifulsoup4==4.6.0
PyGithub==1.35
